break apart collection promise and db promise tests
