+++ test/integration/crud/bulk.test.js
@@ -10,6 +10,7 @@ const {
 const { Long, MongoBatchReExecutionError, MongoDriverError } = require('../../../src');
 const crypto = require('crypto');
 const chai = require('chai');
+const { format: f } = require('util');
 
 const expect = chai.expect;
 chai.use(require('chai-subset'));
@@ -19,6 +20,116 @@ const MAX_BSON_SIZE = 16777216;
 describe('Bulk', function () {
   before(function () {
     return setupDatabase(this.configuration);
+  });
+
+  context('promise tests', () => {
+    it('Should correctly execute unordered bulk operation in promise form', {
+      metadata: {
+        requires: {
+          topology: ['single', 'replicaset', 'sharded', 'ssl', 'heap', 'wiredtiger']
+        }
+      },
+
+      test: function (done) {
+        var configuration = this.configuration;
+        var url = configuration.url();
+        url =
+          url.indexOf('?') !== -1
+            ? f('%s&%s', url, 'maxPoolSize=100')
+            : f('%s?%s', url, 'maxPoolSize=100');
+
+        const client = configuration.newClient(url);
+        client.connect().then(function (client) {
+          var db = client.db(configuration.db);
+          var bulk = db
+            .collection('unordered_bulk_promise_form')
+            .initializeUnorderedBulkOp({ writeConcern: { w: 1 } });
+          bulk.insert({ a: 1 });
+          return bulk
+            .execute()
+            .then(function (r) {
+              test.ok(r);
+              test.deepEqual({ w: 1 }, bulk.s.writeConcern);
+
+              client.close(done);
+            })
+            .catch(done);
+        });
+      }
+    });
+
+    it('Should correctly execute ordered bulk operation in promise form', {
+      metadata: {
+        requires: {
+          topology: ['single', 'replicaset', 'sharded', 'ssl', 'heap', 'wiredtiger']
+        }
+      },
+
+      test: function (done) {
+        var configuration = this.configuration;
+        var url = configuration.url();
+        url =
+          url.indexOf('?') !== -1
+            ? f('%s&%s', url, 'maxPoolSize=100')
+            : f('%s?%s', url, 'maxPoolSize=100');
+
+        const client = configuration.newClient(url);
+        client.connect().then(function (client) {
+          var db = client.db(configuration.db);
+          var bulk = db
+            .collection('unordered_bulk_promise_form')
+            .initializeOrderedBulkOp({ writeConcern: { w: 1 } });
+          bulk.insert({ a: 1 });
+          return bulk
+            .execute()
+            .then(function (r) {
+              test.ok(r);
+              test.deepEqual({ w: 1 }, bulk.s.writeConcern);
+
+              client.close(done);
+            })
+            .catch(done);
+        });
+      }
+    });
+
+    it('Should correctly handle bulkWrite with no options', {
+      metadata: {
+        requires: {
+          topology: ['single', 'replicaset', 'sharded', 'ssl', 'heap', 'wiredtiger']
+        }
+      },
+
+      test: function (done) {
+        var configuration = this.configuration;
+        var client = configuration.newClient(configuration.writeConcernMax(), {
+          maxPoolSize: 1
+        });
+        var error = null;
+        var result = null;
+
+        client
+          .connect()
+          .then(function (client) {
+            var db = client.db(configuration.db);
+            // Get the collection
+            var col = db.collection('find_one_and_replace_with_promise_no_option');
+            return col.bulkWrite([{ insertOne: { document: { a: 1 } } }]);
+          })
+          .then(function (r) {
+            result = r;
+          })
+          .catch(function (err) {
+            error = err;
+          })
+          .then(function () {
+            expect(error).to.not.exist;
+            test.ok(result != null);
+
+            client.close(done);
+          });
+      }
+    });
   });
 
   it('should correctly handle ordered single batch api write command error', {
